Favorito funcionando back

+++ workspace/NgmAdota/src/main/java/com/example/NgmAdota/modules/formInteresse/InteresseModel.java
@@ -23,7 +23,7 @@ public class InteresseModel {
         private UsuarioModel usuarioModel;
 
         @Column(name = "usuario_id", nullable = false)
-        private Integer usuarioId;
+        private Long usuarioId;
 
         @OneToOne()
         @JoinColumn(name = "animal_id", insertable = false, updatable = false)
